[compiler] Mark more nodes as capabilities.
Since V8 uses IntPtrT and UintPtrT to represent both integers and
pointers (and notionally wants to reduce it to a machine word), we add a
boolean flag into a compiler node which represents if the node holds a
pointer, which will be based on a CHERI pointer tag. Some operations
will always return a tagged node (e.g. IntPtrConstant or AllocateRaw),
while others (e.g. IntPtrAdd) will depend on its inputs.

This makes a number of CSA tests pass (namely the ones that try to
allocate objects via CSA). It also fixes IntPtrT loop variables in order
to not generate instruction sequences like:
  ... [c3]
  add x3, x3, #0x10
  ... [c3]
in loops.