[compiler] Tagged -> Word bitcasts aren't pointers.
When bitcasting tagged to a word on CHERI systems, don't mark it as a
pointer representation. Instead, mark it as a word.

Currently, we always do it as a 64-bit word since we don't have V8
working on any 32-bit CHERI systems, but this might need to be revised
for that.