[codegen] Simplify AddSubMacro for CHERI.
This commit cleans up the AddSubMacro to not have so many special cases
and to make use of the existing codegen capabilities better. This should
result in reduced register pressure as well as reduce the number of
instructions needed to perform certain operations, hopefully allowing us
to fit more things in the icache.

Firstly, we don't need to always acquire a temporary register for
AddSub, it's only really necessary when we are operating on the stack
pointer because we can't use scvalue with it.

Further, the shifted register issue has to do with the missing add/sub
instructions in Morello for capability registers. Simply use the X
registers and scvalue afterwards.