[compiler] Mark more nodes as capabilities.
When a TNode is constructed from a CheckedNode, if it's constructing a
RawPtrT or any tagged type, mark it as a capability. We do not do this
for IntPtrT and UintPtrT because they're often used as actual integers
and will require manual handling.

This makes it possible to correctly mark CSA code such as:
  Parameter<Object>(1)

as a capability. Unchecked casts don't work yet.