[CSA] Further propagate capability markings.
This commit touches a lot of bits of V8 in order to mark CSA nodes as
capabilities so that we end up in correct paths in the codegen and use
capability registers. Furthermore, add two helpers,
BitcastCapabilityToAddress<A, P> and BitcastAddressToCapability<P, A>
which can be used to forcefully cast CSA nodes from and to capabilities,
which can be useful in code that misuses IntPtrT, UintPtrT or WordT.

Finally, add helper structs for all of this in order to build up an
is_capability<T>::value and is_capability<T>::maybe_tagged, allowing us
to perform a number of constexpr checks for marking capabilities.