[base] Use CHERI checks in AddressRegion::contains.
By doing it this way, we are able to avoid situations where we attempt
to access constants via the isolate root register (kRootRegister) at
runtime (via CSA) that we can't actually access. An example of this
are isolate->debug()->... pointers, as they aren't actually in the range
of the isolate root capability. Even though they are immovable, and
immutable, we can't access them without going out of bounds of the
isolate root capability, so we need to generate a constant pool in order
to access them. This change ensures that we end up down the correct path
in the compiler (and eventually assembler).