[compiler] Support kCapability64 for representation change.
We can end up in this situation when we have a use representation of
kTagged/kTaggedPointer/kTaggedSigned, but our input representation is a
kCapability64. This commit simply returns the node itself, as that's
what we would expect on the uncompressed build. This *might* be broken
in the compressed configuration, but we have no good way to test that
for now.