[codegen] Accept padreg in push/pop with C-registers.
This is necessary because the baseline assembler has a few helper
functions with variadic template arguments that have no information on
what kinds of registers are being forwarded, and padreg is used by
default. We don't have a good way to check for the register type there,
so instead we accept padreg in the macro assembler's push/pop functions
when used with C registers.