[compiler] Align to pointer size during memory lowering.
When performing reduction of an AllocateRaw, a non-foldable path will
attempt to align the new top to be the expected alignment. On regular
V8, this is only used for 8GB heaps. However, on CHERI, we want to
maintain an alignment to pointer/capability size. This commit ensures
that we always store an aligned new_top in this code path.