[torque] Rework capability identification.
This gives us more control on how each of the types should be identified
as a capability. Furthermore, it makes the rules for a union being
identified as a capability in line with what it should be. This change
should, however, be a no-op in the current implementation, but might be
important in the future.