[torque] Add a --trace-cheri and hook it up to GN.
This commit adds a new flag to the torque compiler, --trace-cheri, which
will trace all the CHERI related events that happen in the Torque
compiler. Currently, this flag only causes a dump of all the struct and
class information, including all of their member names, types, offsets,
any added CHERI-related padding and tracks the progression of the value
of kHeaderSize for each class.

This has been helpful when debugging a couple of Torque problems, so we
keep it in the codebase.