[objects] Rework string internalization on CHERI.
This is in preparation for a Torque change where we will stop aligning
any indexed members to capability alignment. This kind of alignment is
not necessary, but unfortunately without it, we break ThinString-based
internalization of strings because the code expects to either shrink the
allocation or leave it alone, but it cannot grow it. If that is the
case, go via the string forwarding table instead of using ThinStrings.