[objects] Support not being able to externalize strings.
On CHERI systems, we need to insert extra padding to correctly align the
resource pointer in ExternalString. Because of that, we can't always
externalize small strings because their size might be smaller than we
expect it to be. As a result, we can't DCHECK for that property and
instead have to simply return false when it isn't satisfied.

This also has a ripple effect in the tests, so account for it in those
too.