[codegen] Slightly rework loads/stores on CHERI.
A part of this commit is simply running the code through clang-format,
which was not done before. However, the intended functional change here
is to correctly calculate the size which we need to shift a scaled
immediate by. Since in the current version of V8, LDR_c and LDRB_w
conflict in terms of their representations, we can't just use the opcode
to determine the shift size, and instead need to know if it's a
capability or not.

This issue would manifest itself in a fairly unfortunate way, e.g. if
the correct instruction was
  ldrb w5, [c4, #16]
it would generate instead
  ldrb w5, [c4, #1]
and we would now be working with the wrong value in w5, usually crashing
far down the line with a CSA_DCHECK (usually because an array index was
out of bounds, but it was causing other issues too).

This whole backend will at some point need to be reworked (or perhaps
scrapped in favour of a new version of CHERI for arm64 at some point in
the future).