[builtins] Slightly rework deopt.
Firstly, correctly calculate the offset to the space on the stack that
the double registers are stored at. Since on CHERI, we save the
general-purpose registers as C registers, we need to account for their
size when calculating the offset for the double registers.

Secondly, fp_to_sp is an integer, so don't use a capability register to
represent it.