[objects] Iterate pointers from an aligned address.
On CHERI, we can't possibly have a valid pointer on an unaligned
address. There are code paths that call into this function with a
start_offset that doesn't result in an aligned start address because V8
assumes objects to be packed. However, since our allocators
unconditionally align to pointer alignment, and because our Torque
compiler will correctly pad out the fields if needed, we can safely
round up the start_offset here to a pointer alignment and guarantee
that we have a pointer there.