[heap] Ensure blocks are copied in pointer-aligned chunks.
When calling CopyBlock(), we need to make sure that the size in bytes is
aligned to the tagged size, which on uncompressed builds will be
pointer-sized (16 bytes on CHERI). We first copy the bytes 1 by 1 until
we reach a kTaggedSize boundary, at which point we call CopyTagged on
the rest.