[codegen/compiler] Rework Adds.
Adds in V8 is only ever used for actual integers, so even though due to
misuse of IntPtrT and UintPtrT we can end up with capability registers
as inputs and outputs of adds, we don't actually care because they will
only ever be caused by Smis. Furthermore, the old code was just plain
wrong, as it was trying to generate an extended add for a shift of 0,
which would result in something along the lines of
  add c3, c1, w2 sxtw
which is largely nonsensical, as the 9/10 times the inteded code is
actually
  adds x3, x1, x2
and using w2 would lose the top 32 bits and not set anything, producing
nonsensical results.

This commit also reformats the code.