[objects] Pad out base strings on CHERI uncompressed.
The reason we need to pad out the strings manually is because of the
classes that inherit from the base String class. ThinString and
ExternalString both require padding for their first new member to be
aligned as it is a capability (ExternalPointer and String, which starts
with a Map).

This results in cases at runtime, particularly around string tables and
garbage collection, where we need to either internalize or externalize
strings, but cannot do so because the ThinString/ExternalString size is
too large for the string we are trying to internalize/externalize.

This commit also reworks api-callbacks.tq to only use padding on CHERI
uncompressed rather than unconditionally on CHERI systems. We don't need
this padding with pointer compression.