[compiler] Add a CapabilityIsTagged opcode.
In order to have sensible DCHECKs inside Torque and CSA code around
capabilities being tagged, we need a new opcode, kCapabilityIsTagged.
This opcode has 1 value input and 1 value output edge, no effect edges
and no control edges. On Morello, it performs a simple gctag operation,
which gets treated as a boolean in CSA and Torque.

On plain arm64, it simply stores 1 to the output register and completely
discards the input, as it's irrelevant. This mimics the notion that on a
flat memory architecture, everything is tagged and the opcode should not
be used for constructs such as:
  dcheck(!CapabilityIsTagged(value));
without an appropriate @if(CHERI_PURECAP) in Torque or ifdef in CSA.