[compiler] Make GCUnsafeReferenceToRawPtr return tagged capabilities.
There are code paths where the Object we pass in is in fact a
TaggedZeroPattern, while the offset is in fact a tagged capability.
Unfortunately, there doesn't seem to be an easy way to statically
determine that without loads of refactoring, so instead create a branch
that determines which one of them is tagged and then simply propagate
the tag out from that one.

This commit also fixes the fact that IntPtrSub was not generating CapSub
correctly when needed.