[compiler] Only use C registers for stack slots.
Since stack slots work in increments of kSystemPointerSize, we shouldn't
be using X registers (which are 8 bytes) to poke to the stack. Doing
this can result in runtime calls misinterpreting the stack slots.

An example of this is:

  stp x1, x2, [csp, #0x10]
  stp x3, x4, [csp, #0x30]
  br ...

The function we branch to (especially a runtime function) would assume
that c1 is at 0x10, c2 is at 0x20, c3 is at 0x30 and c4 is at 0x40.
However, the reality would be that runtime's "c1" spans both x1 and x2,
followed by garbage data, then runtime's "c3" which spans x3 and x4,
followed by more garbage data as runtime's "c4".