[codegen] Refactor the Morello codegen.
Parts of the codegen would try to take into account that C registers
would end up being passed in, however they implemented the operation in
the wrong way. Furthermore, these code paths never actually get called
with C registers to begin with, so we can delete the code outright and
replace it with simple DCHECKs that the registers being passed in are
not capability registers.

Furthermore, we were using Register::code() to compare the two
registers, but we can just make use of the existing function
AreAliased() for consistency.

Finally, add DCHECKs to ensure that the temporary registers that we need
to acquire during codegen aren't aliased to the register which acts as
the source of the capability provenance.