[heap] Rework finding pages for instruction streams.
When we get an inner_pointer that is a sentry, we need to re-derive the
page from somewhere because modifying the sentry will result in the tag
being stripped. In this commit, we add a new FindPageInSpace method
which will iterate through the list of pages in a space and attempt to
find a pointer which can be used to re-derive the capability.

This approach is far from ideal, though, because it has a very loose
check on what is acceptable as a base capability (we could be out of
bounds), as well as implementing a method which already sort of exists
for large spaces. In the future, it would be good to rework this.