[codegen] Use capability add for shifted immediate operands.
Much like the previous commit, we now have a better understanding of
which code paths lead where into the backend. This commit removes
unnecessary code from the assembler and adds a condition in the macro
assembler for AddSubMacro to generate extended capability add
instructions in the case where an immediate ends up in a shifted
register.