[codegen] Fix codegen for immediate operands.
The previous code could not handle things like:
  Sub(csp, csp, size);

In this commit, we change AddSubMacro to support immediate operands that
cannot be encoded in a single instruction and end up in a register and
want to use capability registers. We have three different cases:
 (1) We can encode the instruction as a capability extended add;
 (2) We have aliased registers and we need to acquire a temporary
     register (e.g. sub csp, csp, x4);
 (3) We can't encode it as a Morello instruction, but our registers
     aren't aliased so we can just use rd as a temporary register.

We also need to change the op to not use the capability instruction in
the last two cases in order to correctly encode the instructions.