[codegen] Use capability memoperands for atomics.
We were tripping up DCHECKs when trying to perform atomic operations
because the memory operand was a capability. This commit loosens the
DCHECKs in atomics to accept capability registers and forcefully casts
ones that need to be capabilities. This isn't strictly necessary in the
current code base, but serves as a way to be explicit and somewhat
future-proof the implementation when merging forward.