[compiler] Don't lose tags in InstructionOperands.
Specific kinds of InstructionOperands, PendingOperands use the value_
field to store a "next" pointer in the form of a linked list. There are
a number of problems with this approach:
 (1) the value_ field is a uint64_t bitmap;
 (2) the way pointers are encoded in the bitmap involves shifting to the
     right and then ORing other stuff on the top (invalidating the
     pointer in the process);
 (3) decoding does the reverse of the above.

There doesn't seem to be an easy, portable way to do this on purecap in
order to preserve tags due to the way capabilities are encoded, as well
as having to make the 'value_' field a uintptr_t, which creates problems
when used as a bitfield in other cases.

Instead, for now just add another field called value_intptr_ which will
only really be used in PendingOperands to maintain a linked list.