[heap] Work around finding untagged objects.
These objects _might_ arise from the gaps that our purecap allocators
create when aligning the objects, but I haven't really confirmed this.
However, doing this seems to make more JS work which is important in the
short-term. It might, however, cause some odd behaviour in the future
and we should be aware of that.