[atomics] Rework AtomicWord for purecap.
Previously, we were treating AtomicWord as an Atomic64 on purecap, which
required some manual intervention around replacing AtomicWord with
AtomicIntPtr. This commit reworks that relationship -- namely it treats
AtomicWord as an AtomicIntPtr as opposed to an Atomic64 on purecap,
which is in line with how the rest of V8 treats Word types.

This has helped remove some duplicated code, but required duplication in
other places. It has also fixed some edge cases around atomics,
particularly around Relaxed_Memcpy and Relaxed_Memmove.