[codegen] Scrap the weird re-derivation of sentries.
This was never really correct and was always a work-around. It also
assumed we would be generating sentries in V8 in a specific way -- which
does not yet exist, nor does it seem like that approach will be used.
Instead, when calling PrepareC64Jump to tag the jump site with the C64
bit (which we may scrap eventually as well), skip sentries all together
because they can no longer be generated by V8 and can only come from
rtld, and those will always have the C64 bit set when we compile in
purecap mode.