[compiler] Use a Word64 for kAllocate during lowering.
Instead of using a Word(), which is defined as a pointer representation,
use an explicit Word64 on purecap. This ensures that in later passes, we
will correctly translate a potential kTaggedSigned representation into a
usable integer. Without this, we end up with a cast from kTaggedSigned
to kCapability64, and due to incorrect uses of kTaggedSigned in various
parts of V8 (i.e. it not being used as an Smi), we can't reliably say
that it's an Smi at that point during the pass.