[cheri] Make LLVM-15 happy.
Since Morello LLVM 14 made everything in a .text section a sentry, we
had to re-derive the sentries from the PCC when deserializing the
embedded data. LLVM 15 fixes this and we get a read-only capability as
opposed to an executable-only sentry. However, this breaks our current
code. For now, work around the issue in order to maintain compatibility
with LLVM 14 until Chromium can switch to LLVM 15. After that, we will
produce our own sentries.