[builtins] Use correct pointer provenance in SubString.
When we call into SubString, in particular the external pointer branch
of it, we will attempt to derive the resulting capability from an offset
as opposed to the real pointer. This will result in an instruction
similar to:
  ldrb w10, [c0, x10]

being generated, where the pointer is actually in $x10 as opposed to
$c0, which is the offset.

This commit generates a runtime check to figure out which of the fields
is tagged (in the CHERI sense) and branches to the code that will
correctly derive the new capability.