[builtins] Ensure we don't lose tags during RecordWriteSaveFP.
We previously forced the bitmap to be an integer under the assumption
that it won't later be used as a pointer. However, the value leaves the
function via the cell pointer, which is in fact expected to be a tagged
capability later on.

Also mark the slot set address as a capability.