[heap] Skip padding objects when iterating stacks.
When we iterate the stack, we can encounter valid objects whose maps are
parts of a kZapValue. Those objects are used for padding and need to be
skipped by the iteration.

This is only a partial workaround of the issue, as the objects can still
be encountered during serialisation and deserialisation. This will be
addressed in another commit later on.