[codegen] Use correct pointer provenance for strings.
Certain string builtins such as StringPrototypeTrim,
StringPrototypeEndsWith and so on can consume strings that have the
actual string pointer in the offset field. Furthermore, they will modify
it using addition/subtraction and create further subslices of those
strings. As a result, the subslices end up with the pointer in the
offset field as well.

Then, references can be created from those subslices, and the pointer
can end up in the start field and the end field instead of the object as
well.

This commit ensures that we generate capability instructions to preserve
tags throughout the offset/start/end fields of those data structures and
modifies LoadReference and StoreReference in CSA to check which value is
tagged and perform a load/store using the correct base capability
register.