[codegen] Treat BigInt as an integer.
Treating BigInt as a capability causes us to generate instructions which
attempt to load a capability from an unaligned offset, since a BigInt's
size is 8. They will never be valid capabilities, so instead of trying
to grow a BigInt to capability size, simply treat them as integers.

Further changes are needed in TurboFan to treat them correctly in the
passes.