[heap] Skip uninitialised objects in stack scanning.
When scanning the stack, we can encounter valid objects that point to
uninitialised maps. We don't actually want to try and get the size of
those objects, as the size is simply defined to be kTaggedSize by the
design of our allocators, so we can just skip over them.

We also don't need to verify them, because these objects are both
expected to appear and only act as padding.