[objects] Relax DCHECKs.
These were causing crashes on valid values being passed in (e.g. Smis)
which are checked for later on. We only want to check for alignment if
the value is tagged. If it's not, and it should be, we will just crash
attempting to use it and should still get a valid backtrace.