[compiler] Fix kCapability64 representation changes.
The main issue surfaced when kCapability64 needed to be converted to a
kTaggedSigned (Smi) representation. This would result in missing
lsl ..., #32 instructions and would cause crashes at runtime. This
commit mostly changes the way kCapability64 representations are treated
to match kWord64. The most notable exception to this is when a
kCapability64 is trying to be converted to some kind of a pointer, which
kWord64 representations don't support. In that case, we simply return
the current node, as a capability is already a pointer in those cases.