[cheri] [ptr-compr] Initial version that builds.
This commit removes the distinction between the is_capability<T> helper
for compressed and uncompressed build as an initial experiment in
porting pointer compression.

Furthermore, it ensures that a correct register type is used in a
builtin when pointer compression is configured and modifies two DCHECKs
to match the expectations with pointer compression enabled.