[objects] Change the size of kInstructionStartOffset.
There is no need for it to be sized kSystemPointerSize + 3 * kTaggedSize
for an uncompressed pointer build. Additionally, only perform aligned
writes and reads when pointer compression is enabled, as we should never
need it without it. Further, ensure the code object is aligned to
pointer size before we access the kInstructionStart offset as a
workaround for an unaligned load. This should probably be fixed
differently, but it gets us further for now.

Finally, this commit also replaces __builtin_* use with V8-specific
macros.