[compiler] Store the data_ptr at an aligned location.
When compiling a typed array, we need to store its data_ptr argument at
an aligned location, which is not sizeof(size_t), but sizeof(uintptr_t).
Furthermore, a similar change has to be made for the length in another
place in order to not overwrite 8 bytes of the data_ptr after storing
it.

This commit also adds a helper static_assert for alignment of
FastApiCallbackOptions in wasm_memory.