[compiler] Work around VisitUnalignedLoad being called.
Currently, we just call VisitLoad. This isn't the same as saying that
everything can be accessed unaligned, however, because other checks do
exists. For now, we just work around the issue by calling VisitLoad.
Ideally, we would have some extra debugging helpers here if V8 is
compiled in debug mode.

Note that we can't really UnalignedLoad a capability no matter what
without reconstructing it from somewhere, and we shouldn't be doing
that.