[builtins] Correctly compute the float section for WASM.
Even though kSystemPointerSize is going to be 16 bytes on CHERI, the
rest of the code assumes that we're shifting by
kSystemPointerSizeLog2 + 1 rather than just kSystemPointerSizeLog2. This
results in loading the float section from the incorrect memory location,
breaking WASM <-> JS interaction.

We currently address this by encoding two identical Sub instructions
because we can't encode a shift by 5 on Morello without needing
additional instructions (e.g. acquire a temporary register to perform
the shift on and then scvalue the csp). This might not be the right
thing to do long-term and should be revisited at some point.