[torque] Fix incompatibility of a CHERI change.
This issue manifested in the form that too much padding was added by
just having V8_SFI_NEEDS_PADDING as a condition. Instead, the amount of
padding added on CHERI and non-CHERI should be different, so we added
two more defines into Torque to distinguish between those situations.

While this is ugly, it avoids having to add boolean logic to Torque's
proprecossing.