[compiler] [codegen] Fix more build errors on non-CHERI.
TNode::IsCapability() is shimmed, but it always returns false because
nothing on a flat architecture is a capability. For the cases where we
want to DCHECK that it in fact is a capability, we need to conditionally
compile it for CHERI.

We might want to, at some point, turn this into a DCHECK_IMPLIES.
However, this will do for now.