[cheri] Make cross-compilation possible.
This commit adds usage of a new build flag, is_cheri, which specifies
that the build of V8 should target a CHERI architecture. It also adds a
new toolchain, //tools/toolchain:morello-buildroot, which can be used as
a helper to cross-compile to CheriBSD-Morello.

Additionally, this commit refactors Torque's use of ifdefs around CHERI
to instead utilise a command-line argument called -cheriabi to signify
that Torque sources should be compiled for CHERI. This flag is
automatically passed in based on the is_cheri gn argument.

This doesn't quite work yet, but it gets us to the point where we *can*
cross-compile V8 for morello. The main missing step appears to be a
refactor of using V8_TARGET_CHERI instead of __CHERI_PURE_CAPABILITY__
so that mksnapshot running on linux/x64 will correctly generate
capability instructions via TurboFan as opposed to regular arm64
instructions.