[cheri] Major refactor of the codebase.
This commit does a whole lot of things that are quite difficult to break
up into self-contained units.

The main parts of this commit are:
 (1) Replacing usage of #ifdef __CHERI_PURE_CAPABILITY__ with #if
     V8_TARGET_CHERI;
 (2) Replacing ifdef'd DCHECKs with a
     DCHECK_IMPLIES(V8_TARGET_CHERI_BOOL, ...);
 (3) Creating shims for all the capability registers (c0-c30, csp, cfp,
     clr, padregc, ...) on non-CHERI builds, allowing us to avoid ifdefs
     in the code;
 (4) Refactors any use of those registers to always use capability
     registers, which will shim to X registers on flat arm64;
 (5) Adds shims for CPURegister::C(), Operand::ToC() and various other
     bits on flat arm64;
 (6) Refactors all the codegen to make use of them, getting rid of
     further ifdefs;
 (7) Refactors turboshaft and turbofan to use fewer ifdefs;

An additional goal of this commit is to support cross-compilation to
CHERI. We cannot do this from a conventional 64-bit architecture due to
pointer size mismatch. However, this refactor sets us up to support
cross-compiling to CHERI from any architecture that may have a matching
pointer size in the future (e.g. cross-compiling to RV32Y from x64).
There are still some issues in the code that will need to be resolved
once this is actually attempted, but the effort should be significantly
smaller than without this commit.