[cheri] Add initial c18n support.
This commit makes the following changes to the general operation of both
V8 and Oilpan:
 - The stack scanning algorithm now assumes it is being called through a
   c18n trampoline (security.cheri.lib_based_c18n_wrap_fptr=1);
 - Instead of using pthread/reading the current csp, it skips the
   current stack, which it can do because determining the stack top and
   bottom happened before this function call and the current stack
   values should have no effect on that;
 - Gets the current stack top and bottom from rtld instead;
 - Scans the stack and iteratively updates the stack top and bottom
   until we run out of stacks.

The above functionality is encapsulated in the OS class and a new C18n
class, and implemented per-platform to support differing underlying
mechanisms for c18n.

Additional changes had to be made to snapshot deserialization due to
function pointers now being wrapped by trampolines, so our sentries
instead point to a trampoline instead of the builtins. This can be done
using OS::C18n::Reflect().

Finally, this commit simplifies some code around re-deriving sentries.