[wasm] Use RelocatableCapability64Constant for a stub call.
Not doing so will generate a RelocatableInt64Constant because the
argument to RelocatableIntPtrConstant is not a capability, and therefore
will be untagged. However, this will generate an instruction stream that
will eventually try to branch to this constant, which *will* be a valid
capability at that time, but we will be working with a
RelocatableInt64Constant rather than a RelocatableCapability64Constant,
tripping a DCHECK or raising an exception when we actually try to branch
to it, as it won't be done using a capability.