[heap] Make possibly empty slot buckets CHERI-compatible.
The original code uses uintptr_t to store information in a bitfield,
causing issues on CHERI. However, we cannot simply replace the usage of
uintptr_t with an integer type, as the bitmap_ field gets promoted to a
pointer to an array if there is no more room to store bits. Work around
this issue by defining a WordType data type inside the class as
ptraddr_t (which is shimmed to uintptr_t on non-purecap builds) and
insert casts, as well as DCHECKs in appropriate locations.

This commit also adds a V8_CHERI_PURECAP_BOOL define which can be used
without using an #ifdef for CHERI-specific code.