[wasm] Workaround: Map the initial code region to 1GB.
Without this, we cannot merge the surrounding separately mmaped regions.
Ideally, we would have some kind of a mremap call instead when merging
the regions rather than mapping this to begin with, especially since we
don't call subsequent mmaps with MAP_FIXED at the moment. However, it
appears to work fine.