[compiler] Ensure the pending message object is tagged.
When we store/load the pending message object, we need to do so using
capability instructions. The issue in the existing code was using the
IntPtr MachineType, which amounts to Int64/Int32 depending on the system
pointer size. Instead, we use a newly defined access builder method
ForExternalPointer() which ensures we will use capability instructions
in the codegen.